refactor(mapgen): consolidate operation authoring
