+++ docs/system/libs/mapgen/explanation/DOMAIN-MODELING.md
@@ -144,7 +144,7 @@ Preferred dependency direction:
 
 - Domain contract authoring: `packages/mapgen-core/src/authoring/domain/contract.ts`
 - Domain router authoring: `packages/mapgen-core/src/authoring/domain/router.ts`
-- Op authoring helpers: `packages/mapgen-core/src/authoring/op/contract.ts`
+- Op authoring helpers: `packages/mapgen-core/src/authoring/operation/contract.ts`
 - Example aggregate contract/router: `mods/mod-swooper-maps/src/domain/foundation/contract.ts`, `mods/mod-swooper-maps/src/domain/foundation/router.ts`
 - Example direct module: `mods/mod-swooper-maps/src/domain/foundation/modules/tectonics/`
 - Domain contract index: `docs/system/libs/mapgen/reference/domains/DOMAINS.md`
