+++ docs/system/libs/mapgen/how-to/add-an-op.md
@@ -168,8 +168,8 @@ const implementations = {
 
 ## Ground truth anchors
 
-- Op contract API: `packages/mapgen-core/src/authoring/op/contract.ts`
-- Op implementation wrapper: `packages/mapgen-core/src/authoring/op/create.ts`
+- Op contract API: `packages/mapgen-core/src/authoring/operation/contract.ts`
+- Op implementation wrapper: `packages/mapgen-core/src/authoring/operation/create.ts`
 - Domain registry authoring: `packages/mapgen-core/src/authoring/domain/contract.ts`
 - Example op contract: `mods/mod-swooper-maps/src/domain/foundation/modules/tectonics/ops/compute-era-tectonic-fields/contract.ts`
 - Example op implementation: `mods/mod-swooper-maps/src/domain/foundation/modules/tectonics/ops/compute-era-tectonic-fields/index.ts`
