+++ packages/mapgen-core/src/authoring/domain/authority.ts
@@ -1,4 +1,4 @@
-import type { OpContractAny } from "../op/contract.js";
+import type { OpContractAny } from "../operation/contract.js";
 import type { OwnDataRecord } from "../own-data-record.js";
 
 type OperationContracts = OwnDataRecord<OpContractAny>;
