+++ packages/mapgen-core/src/authoring/operation/contract.ts
