+++ packages/mapgen-core/src/authoring/operation/create.ts
