+++ packages/mapgen-core/src/authoring/operation/envelope.ts
