+++ packages/mapgen-core/src/authoring/operation/input-admission.ts
