+++ packages/mapgen-core/src/authoring/operation/strategy-definition.ts
