+++ packages/mapgen-core/src/authoring/operation/types.ts
