+++ packages/mapgen-core/src/authoring/step/contract.ts
@@ -3,9 +3,9 @@ import type { DependencyTag } from "@mapgen/engine/index.js";
 import { type TObject, type TSchema, Type } from "typebox";
 import { type Artifact, assertArtifact } from "../artifact/contract.js";
 import { freezeContractGraph, snapshotContractGraph } from "../contract-graph.js";
-import { isCanonicalOpContract } from "../op/contract.js";
-import { buildOpEnvelopeSchema } from "../op/envelope.js";
-import type { OpTypeBagOf } from "../op/types.js";
+import { isCanonicalOpContract } from "../operation/contract.js";
+import { buildOpEnvelopeSchema } from "../operation/envelope.js";
+import type { OpTypeBagOf } from "../operation/types.js";
 import { applySchemaConventions } from "../schema.js";
 import { registerCanonicalStepContractInternal } from "./authority.js";
 import { assertNoStepStageIdentityAliases } from "./identity.js";
